+++ Cities/City_List.cs
@@ -5,12 +5,12 @@ namespace City
 {
     public abstract class City_List
     {
-        static Dictionary<uint, City_Data> _defaultCities;
-        public static Dictionary<uint, City_Data> DefaultCities => _defaultCities ??= _initialiseDefaultCities();
+        static Dictionary<ulong, City_Data> _defaultCities;
+        public static Dictionary<ulong, City_Data> DefaultCities => _defaultCities ??= _initialiseDefaultCities();
         
-        static Dictionary<uint, City_Data> _initialiseDefaultCities()
+        static Dictionary<ulong, City_Data> _initialiseDefaultCities()
         {
-            return new Dictionary<uint, City_Data>
+            return new Dictionary<ulong, City_Data>
             {
                 {
                     1, new City_Data
@@ -20,12 +20,12 @@ namespace City
                         cityDescription: "City 1 Description",
                         cityFactionID: 1,
                         regionID: 1,
-                        allJobSiteIDs: new List<uint>
+                        allJobSiteIDs: new List<ulong>
                         {
                             1
                         },
                         population: new PopulationData(
-                            allCitizenIDList: new List<uint>(),
+                            allCitizenIDList: new List<ulong>(),
                             maxPopulation: 100,
                             expectedPopulation: 50),
                         prosperityData: new ProsperityData(
